g-eagle/nlq_agent/requirements.txt
# NLQ Agent — Python Dependencies

# ── LangChain + LangGraph ──
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# ── Vector DB ──
pinecone-client>=3.0.0

# ── AWS ──
boto3>=1.34.0

# ── Config ──
pyyaml>=6.0
python-dotenv>=1.0.0

# ── HTTP (for DataHub) ──
requests>=2.31.0

# ── Evaluation (optional) ──
# langsmith>=0.1.0
